# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/players":{"post":{"operationId":"PlayersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePlayerDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"PlayersController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/players/{id}":{"get":{"operationId":"PlayersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"PlayersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePlayerDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"PlayersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/real-games":{"post":{"operationId":"RealGamesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRealGameDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"RealGamesController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/real-games/{id}":{"get":{"operationId":"RealGamesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"RealGamesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRealGameDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"RealGamesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/real-games/{realGameId}/players":{"post":{"operationId":"RealGamePlayersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRealGamePlayerDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"RealGamePlayersController_findAll","parameters":[{"name":"realGameId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/real-games/{realGameId}/players/{id}":{"get":{"operationId":"RealGamePlayersController_findOne","parameters":[{"name":"realGameId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"RealGamePlayersController_update","parameters":[{"name":"realGameId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRealGamePlayerDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"RealGamePlayersController_remove","parameters":[{"name":"realGameId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/teams":{"post":{"operationId":"TeamsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTeamDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"TeamsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/teams/{id}":{"get":{"operationId":"TeamsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"TeamsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTeamDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"TeamsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"BAMG API","description":"Bamp API to interact with players and teams","version":"1.0","contact":{}},"tags":[{"name":"bamg","description":""}],"servers":[],"components":{"schemas":{"CreatePlayerDto":{"type":"object","properties":{"playerOptaId":{"type":"string","description":"Unique identifier for the player from Opta"},"firstName":{"type":"string","description":"First name of the player"},"lastName":{"type":"string","description":"Last name of the player"},"position":{"type":"string","description":"Position of the player"},"optaPosition":{"type":"string","description":"Opta position of the player"},"teamId":{"type":"number","description":"Team ID"},"teamOptaId":{"type":"string","description":"Team Opta ID"},"attack":{"type":"number","description":"Attack skill level"},"defense":{"type":"number","description":"Defense skill level"},"strength":{"type":"number","description":"Strength level"},"impact":{"type":"number","description":"Impact level"},"skills":{"type":"number","description":"Skills level"},"scoring":{"type":"number","description":"Scoring ability"}},"required":["playerOptaId","firstName","lastName","position","optaPosition","teamOptaId","attack","defense","strength","impact","skills","scoring"]},"UpdatePlayerDto":{"type":"object","properties":{"firstName":{"type":"string","description":"First name of the player"},"lastName":{"type":"string","description":"Last name of the player"}},"required":["firstName","lastName"]},"CreateRealGameDto":{"type":"object","properties":{}},"UpdateRealGameDto":{"type":"object","properties":{}},"CreateRealGamePlayerDto":{"type":"object","properties":{}},"UpdateRealGamePlayerDto":{"type":"object","properties":{}},"CreateTeamDto":{"type":"object","properties":{"teamOptaId":{"type":"string","example":"12345","description":"Unique Opta ID for the team"},"name":{"type":"string","example":"Team Name","description":"Name of the team"},"coachOptaId":{"type":"string","example":"67890","description":"Opta ID for coach"}},"required":["teamOptaId","name"]},"UpdateTeamDto":{"type":"object","properties":{"name":{"type":"string","description":"Name of the team"},"coachOptaId":{"type":"string","example":"67890","description":"Opta ID for coach"}},"required":["name"]}}}}